+++ experimental/packages/api-logs/src/index.ts
@@ -9,7 +9,7 @@ export { SeverityNumber } from './types/LogRecord';
 export type { LogAttributes, LogBody, LogRecord } from './types/LogRecord';
 export type { LoggerOptions } from './types/LoggerOptions';
 export type { AnyValue, AnyValueMap } from './types/AnyValue';
-export { NOOP_LOGGER, NoopLogger } from './NoopLogger';
+export { createNoopLogger } from './NoopLogger';
 
 import { LogsAPI } from './api/logs';
 export const logs = LogsAPI.getInstance();