+++ experimental/packages/api-logs/test/noop-implementations/noop-logger.test.ts
@@ -5,15 +5,9 @@
 
 import * as assert from 'assert';
 import { SeverityNumber } from '../../src';
-import { NoopLogger } from '../../src/NoopLogger';
 import { NoopLoggerProvider } from '../../src/NoopLoggerProvider';
 
 describe('NoopLogger', () => {
-  it('constructor should not crash', () => {
-    const logger = new NoopLoggerProvider().getLogger('test-noop');
-    assert.ok(logger instanceof NoopLogger);
-  });
-
   it('calling emit should not crash', () => {
     const logger = new NoopLoggerProvider().getLogger('test-noop');
     logger.emit({